services/pyproject.toml
[tool.poetry]
name = "engage-api"
version = "0.1.0"
description = "Engage-AI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
httpx = "^0.27.0"
openai = "^1.35.7"
sqlmodel = "^0.0.19"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
pyjwt = "^2.8.0"
msgpack = "^1.0.8"
psycopg2-binary = "^2.9.9"
fastapi-utilities = "^0.2.0"
boto3 = "^1.34.157"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
bandit = "^1.7.9"
safety = "^3.2.3"
alembic = "^1.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
